+++ packages/provider-utils/src/index.ts
@@ -22,6 +22,7 @@ export {
 } from './create-provider-stream-error';
 export * from './delay';
 export { DelayedPromise } from './delayed-promise';
+export * from './delete-from-api';
 export {
   detectMediaType,
   getTopLevelMediaType,
@@ -64,6 +65,7 @@ export { normalizeBatchRequestCounts } from './normalize-batch-request-counts';
 export * from './parse-json';
 export { parseJsonEventStream } from './parse-json-event-stream';
 export { parseProviderOptions } from './parse-provider-options';
+export * from './post-multipart-stream-to-api';
 export * from './post-to-api';
 export {
   createProviderDefinedToolFactory,